DAO refactoring

+++ contracts/DAO/DAOInterface.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.4.11;
 interface DAOInterface {
     function isParticipant(address participantAddress) constant returns (bool);
 
-    function getMinVotes() public constant returns(uint8);
+    function getMinVotes() public constant returns(uint);
 
     function getParticipantsCount() public constant returns(uint);
 }
